refactor: 패키지명 변경
- global 패키지를 _global로 변경하였다.
