+++ starbucks/src/main/java/git_kkalnane/backend/starbucks/_global/success/SuccessResponse.java
@@ -1,4 +1,4 @@
-package git_kkalnane.backend.starbucks.global.success;
+package git_kkalnane.backend.starbucks._global.success;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
