+++ starbucks/src/main/java/git_kkalnane/backend/starbucks/auth/common/exception/AuthException.java
@@ -1,7 +1,7 @@
 package git_kkalnane.backend.starbucks.auth.common.exception;
 
 
-import git_kkalnane.backend.starbucks.global.error.core.BaseException;
+import git_kkalnane.backend.starbucks._global.error.core.BaseException;
 
 public class AuthException extends BaseException {
