+++ starbucks/src/main/java/git_kkalnane/backend/starbucks/auth/common/jwt/JwtTokenProvider.java
@@ -3,7 +3,7 @@ package git_kkalnane.backend.starbucks.auth.common.jwt;
 
 import git_kkalnane.backend.starbucks.auth.common.exception.AuthErrorCode;
 import git_kkalnane.backend.starbucks.auth.common.exception.AuthException;
-import git_kkalnane.backend.starbucks.global.utils.GlobalLogger;
+import git_kkalnane.backend.starbucks._global.utils.GlobalLogger;
 import io.jsonwebtoken.*;
 import io.jsonwebtoken.security.Keys;
 import org.springframework.stereotype.Component;
