+++ starbucks/src/main/java/git_kkalnane/backend/starbucks/member/common/exception/MemberException.java
@@ -1,6 +1,6 @@
 package git_kkalnane.backend.starbucks.member.common.exception;
 
-import git_kkalnane.backend.starbucks.global.error.core.BaseException;
+import git_kkalnane.backend.starbucks._global.error.core.BaseException;
 
 public class MemberException extends BaseException {
     public MemberException(MemberErrorCode errorCode) {
