+++ starbucks/src/main/java/git_kkalnane/backend/starbucks/member/domain/Member.java
@@ -1,6 +1,6 @@
 package git_kkalnane.backend.starbucks.member.domain;
 
-import git_kkalnane.backend.starbucks.global.entity.BaseTimeEntity;
+import git_kkalnane.backend.starbucks._global.entity.BaseTimeEntity;
 import git_kkalnane.backend.starbucks.member.dto.request.SignUpRequest;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
