+++ starbucks/src/main/java/git_kkalnane/backend/starbucks/merchant/domain/Merchant.java
@@ -1,6 +1,6 @@
 package git_kkalnane.backend.starbucks.merchant.domain;
 
-import git_kkalnane.backend.starbucks.global.entity.BaseTimeEntity;
+import git_kkalnane.backend.starbucks._global.entity.BaseTimeEntity;
 import jakarta.persistence.*;
 import lombok.*;
